SDP | Centralize folder and constants.

+++ authentication/services/removeUserRegisterService.go
@@ -4,6 +4,7 @@ import (
 	model "Go_Watchlist/userAuthentication/models"
 	repo "Go_Watchlist/userAuthentication/repos"
 	// "regexp"
+	"errors"
 )
 
 // func RegisterUser(req model.AddUserRequest) error {
@@ -26,6 +27,16 @@ import (
 // 	return repo.InsertUser(dbUser)
 // }
 
+
 func RemoveRegisterUser(req model.RemoveUserRequest) error {
+	exists, err := repo.UserExistsByPANAndMobile(req.PAN, req.Mobile)
+	if err != nil {
+		return err
+	}
+	if !exists {
+		return errors.New("user with given PAN and mobile does not exist or mismatch")
+	}
 	return repo.RemoveUserByPAN(req.PAN)
 }
+
+
